It _does_ work!
The fix was: don't repeat the request after seeing "pending."  Turns out
the cert was actually being issued.

Besides that, a "--staging" option was added to allow testing against
Let's Encrypt staging servers.  This is necessary for success because I
am now rate-limited!

Error reporting is improved and code is cleaned up somewhat.